+++ src/main/java/com/github/theredbrain/inventorysizeattributes/InventorySizeAttributes.java
@@ -1,19 +1,11 @@
 package com.github.theredbrain.inventorysizeattributes;
 
 import com.github.theredbrain.inventorysizeattributes.config.ServerConfig;
-import com.github.theredbrain.inventorysizeattributes.config.ServerConfigWrapper;
-import com.google.gson.Gson;
-import me.shedaniel.autoconfig.AutoConfig;
-import me.shedaniel.autoconfig.serializer.JanksonConfigSerializer;
-import me.shedaniel.autoconfig.serializer.PartitioningSerializer;
+import com.github.theredbrain.inventorysizeattributes.entity.player.DuckPlayerEntityMixin;
+import me.fzzyhmstrs.fzzy_config.api.ConfigApiJava;
 import net.fabricmc.api.ModInitializer;
-import net.fabricmc.fabric.api.networking.v1.PayloadTypeRegistry;
-import net.fabricmc.fabric.api.networking.v1.ServerPlayConnectionEvents;
-import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
 import net.minecraft.entity.attribute.EntityAttribute;
-import net.minecraft.network.RegistryByteBuf;
-import net.minecraft.network.codec.PacketCodec;
-import net.minecraft.network.packet.CustomPayload;
+import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.registry.entry.RegistryEntry;
 import net.minecraft.util.Identifier;
 import org.slf4j.Logger;
@@ -23,42 +15,22 @@ import org.slf4j.LoggerFactory;
 public class InventorySizeAttributes implements ModInitializer {
 	public static final String MOD_ID = "inventorysizeattributes";
 	public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
-	public static ServerConfig serverConfig;
+	public static ServerConfig SERVER_CONFIG = ConfigApiJava.registerAndLoadConfig(ServerConfig::new);
 
 	public static RegistryEntry<EntityAttribute> HOTBAR_SLOT_AMOUNT;
 	public static RegistryEntry<EntityAttribute> INVENTORY_SLOT_AMOUNT;
 
+	public static int getActiveHotbarSlotAmount(PlayerEntity playerEntity) {
+		return Math.min(9, Math.max(0, (Math.min(9, Math.max(0, InventorySizeAttributes.SERVER_CONFIG.default_hotbar_slot_amount.get())) + ((DuckPlayerEntityMixin)playerEntity).inventorysizeattributes$getHotbarSlotAmount())));
+	}
+
+	public static int getActiveInventorySlotAmount(PlayerEntity playerEntity) {
+		return Math.min(27, Math.max(0, (Math.min(27, Math.max(0, InventorySizeAttributes.SERVER_CONFIG.default_inventory_slot_amount.get())) + ((DuckPlayerEntityMixin)playerEntity).inventorysizeattributes$getInventorySlotAmount())));
+	}
+
 	@Override
 	public void onInitialize() {
 		LOGGER.info("Inventories come in different sizes now!");
-
-		// Config
-		AutoConfig.register(ServerConfigWrapper.class, PartitioningSerializer.wrap(JanksonConfigSerializer::new));
-		serverConfig = ((ServerConfigWrapper) AutoConfig.getConfigHolder(ServerConfigWrapper.class).getConfig()).server;
-
-		PayloadTypeRegistry.playS2C().register(ServerConfigSyncPacket.PACKET_ID, ServerConfigSyncPacket.PACKET_CODEC);
-		ServerPlayConnectionEvents.JOIN.register((handler, sender, server) -> {
-			ServerPlayNetworking.send(handler.player, new ServerConfigSyncPacket(serverConfig));
-		});
-
-	}
-
-	public record ServerConfigSyncPacket(ServerConfig serverConfig) implements CustomPayload {
-		public static final CustomPayload.Id<ServerConfigSyncPacket> PACKET_ID = new CustomPayload.Id<>(identifier("server_config_sync"));
-		public static final PacketCodec<RegistryByteBuf, ServerConfigSyncPacket> PACKET_CODEC = PacketCodec.of(ServerConfigSyncPacket::write, ServerConfigSyncPacket::new);
-
-		public ServerConfigSyncPacket(RegistryByteBuf registryByteBuf) {
-			this(new Gson().fromJson(registryByteBuf.readString(), ServerConfig.class));
-		}
-
-		private void write(RegistryByteBuf registryByteBuf) {
-			registryByteBuf.writeString(new Gson().toJson(serverConfig));
-		}
-
-		@Override
-		public CustomPayload.Id<? extends CustomPayload> getId() {
-			return PACKET_ID;
-		}
 	}
 
 	public static Identifier identifier(String path) {