+++ src/main/java/com/github/theredbrain/inventorysizeattributes/mixin/screen/PlayerScreenHandlerMixin.java
@@ -1,0 +1,33 @@
+package com.github.theredbrain.inventorysizeattributes.mixin.screen;
+
+import com.github.theredbrain.inventorysizeattributes.entity.player.DuckPlayerEntityMixin;
+import com.github.theredbrain.slotcustomizationapi.api.SlotCustomization;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.entity.player.PlayerInventory;
+import net.minecraft.recipe.CraftingRecipe;
+import net.minecraft.recipe.input.CraftingRecipeInput;
+import net.minecraft.screen.AbstractRecipeScreenHandler;
+import net.minecraft.screen.PlayerScreenHandler;
+import net.minecraft.screen.ScreenHandlerType;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(PlayerScreenHandler.class)
+public abstract class PlayerScreenHandlerMixin extends AbstractRecipeScreenHandler<CraftingRecipeInput, CraftingRecipe> {
+
+	public PlayerScreenHandlerMixin(ScreenHandlerType<?> screenHandlerType, int i) {
+		super(screenHandlerType, i);
+	}
+
+	@Inject(method = "<init>(Lnet/minecraft/entity/player/PlayerInventory;ZLnet/minecraft/entity/player/PlayerEntity;)V", at = @At("TAIL"))
+	public void PlayerScreenHandler(PlayerInventory inventory, boolean onServer, PlayerEntity owner, CallbackInfo ci) {
+		for (int i = 36; i < 45; i++) {
+			((SlotCustomization) this.slots.get(i)).slotcustomizationapi$setDisabledOverride(i >= 36 + ((DuckPlayerEntityMixin) owner).inventorysizeattributes$getActiveHotbarSlotAmount());
+		}
+		for (int i = 9; i < 36; i++) {
+			((SlotCustomization) this.slots.get(i)).slotcustomizationapi$setDisabledOverride(i >= 9 + ((DuckPlayerEntityMixin) owner).inventorysizeattributes$getActiveInventorySlotAmount());
+		}
+	}
+}