+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.1.0
+
+- removed dependency on cloth config
+- added dependency on fzzy config
+- added missing localization keys
+- added API methods for getting active hotbar/inventory size of a player to main class
+- inventory size is now also checked when opening the normal inventory
+
 # 1.0.1
 
 - internal refactoring to improve compatibility